refactor: restructure authentication flow by removing unused components and implementing a new layout for email verification
