+++ src/app/(navgroup)/layout.tsx
@@ -1,0 +1,42 @@
+import type { Metadata } from "next";
+import "@/app/globals.css";
+import Navbar from "@/app/(navgroup)/navbar";
+import { cookies } from "next/headers";
+import jwt from "jsonwebtoken";
+
+export const metadata: Metadata = {
+  title: "Create Next App",
+  description: "Generated by create next app",
+};
+
+export default async function NavbarLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  const token = (await cookies()).get("token")?.value || "";
+  let isUserVerified = false;
+
+  try {
+    // The middleware already protects routes, this check is for UI rendering.
+    if (token) {
+      const decoded = jwt.verify(token, process.env.TOKEN_SECRET!) as {
+        isVerified: boolean;
+      };
+      if (decoded) {
+        isUserVerified = decoded.isVerified;
+      }
+    }
+  } catch (error) {
+    // Token is invalid or expired
+    console.log("Invalid token:", error);
+    isUserVerified = false;
+  }
+
+  return (
+    <div className="min-h-screen bg-gradient-to-t from-black to-purple-600">
+      <Navbar isVerified={isUserVerified} />
+      {children}
+    </div>
+  );
+}
