+++ src/app/layout.tsx
@@ -1,9 +1,6 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
-import "./globals.css";
-import Navbar from "@/app/navbar";
-import { cookies } from "next/headers";
-import jwt from "jsonwebtoken";
+import "@/app/globals.css";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
@@ -20,37 +17,15 @@ export const metadata: Metadata = {
   description: "Generated by create next app",
 };
 
-export default async function RootLayout({
+export default function RootLayout({
   children,
 }: Readonly<{
   children: React.ReactNode;
 }>) {
-  const token = (await cookies()).get("token")?.value || "";
-  let isUserVerified = false;
-
-  try {
-    // The middleware already protects routes, this check is for UI rendering.
-    if (token) {
-      const decoded = jwt.verify(token, process.env.TOKEN_SECRET!) as {
-        isVerified: boolean;
-      };
-      if (decoded) {
-        isUserVerified = decoded.isVerified;
-      }
-    }
-  } catch (error) {
-    // Token is invalid or expired
-    console.log("Invalid token:", error);
-    isUserVerified = false;
-  }
-
   return (
     <html lang="en">
-      <body
-        className={`${geistSans.variable} ${geistMono.variable} antialiased`}
-      >
+      <body className={`${geistSans.variable} ${geistMono.variable}`}>
         <div className="min-h-screen bg-gradient-to-t from-black to-purple-600">
-          <Navbar isVerified={isUserVerified} />
           {children}
         </div>
       </body>
